Address AI review round 5: fix bootstrap NaN propagation, fix CI pandas compat
- Guard overall bootstrap stats in _run_bootstrap() when ATT is NaN,
  preventing _compute_bootstrap_pvalue from returning 1/(B+1) instead of NaN
- Add test_no_post_effects_bootstrap_returns_nan for the bootstrap NaN path
- Cast first_treat to float before assigning np.inf in tests for pandas compat

+++ tests/test_staggered.py
@@ -111,8 +111,9 @@ class TestCallawaySantAnna:
             data.copy(), outcome="outcome", unit="unit", time="time", first_treat="first_treat"
         )
 
-        # Re-encode never-treated from 0 to np.inf
+        # Re-encode never-treated from 0 to np.inf (cast to float first for pandas compat)
         data_inf = data.copy()
+        data_inf["first_treat"] = data_inf["first_treat"].astype(float)
         data_inf.loc[data_inf["first_treat"] == 0, "first_treat"] = np.inf
 
         results_inf = cs.fit(
